+++ src/services/sound.cpp
@@ -213,7 +213,7 @@ bool Sound::Update(ISoundSource* /*src*/) {
     // sample = ticks * (rate/50) / clock / 2000
 
     // MulDiv(a, b, c) = (int64) a * b / c
-    int a = MulDiv(time, rate50_, pc_->GetEffectiveSpeed()) + tdiff_;
+    int a = MulDiv(time, rate50_, static_cast<int>(pc_->GetEffectiveSpeed())) + tdiff_;
     //      a = MulDiv(a, mixrate, samplingrate);
     int samples = a / 2000;
     tdiff_ = a % 2000;
